GH-38090: [C++][Emscripten] compute: Suppress shorten-64-to-32 warnings
We need explicit cast to use `int64_t` for `size_t` on Emscripten.

Explicit casts.

+++ cpp/src/arrow/compute/exec.h
@@ -393,7 +393,7 @@ struct ARROW_EXPORT ExecSpan {
   /// \brief Return the value at the i-th index
   template <typename index_type>
   inline const ExecValue& operator[](index_type i) const {
-    return values[i];
+    return values[static_cast<size_t>(i)];
   }
 
   /// \brief A convenience for the number of values / arguments.
